Revert "TopBar, SingleTitleTopBar, SearchTitleTopBar, DoubleTitleTopBar의 머티리얼 의존성 제거"
This reverts commit 51fad9bc105ffab9c3ff746c36e6b75427459f4a.
