+++ compose/src/main/java/com/yourssu/design/system/compose/component/TopBar.kt
@@ -1,4 +1,4 @@
-package com.yourssu.design.system.compose.component.topbar
+package com.yourssu.design.system.compose.component
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.PaddingValues
@@ -8,7 +8,7 @@ import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.wrapContentHeight
+import androidx.compose.material.TopAppBar
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -19,32 +19,48 @@ import com.yourssu.design.system.compose.YdsTheme
 import com.yourssu.design.system.compose.atom.TopBarButton
 import com.yourssu.design.system.compose.base.YdsScaffold
 import com.yourssu.design.system.compose.base.YdsText
-import com.yourssu.design.system.compose.component.TopAppBar
 
 @Composable
-fun SingleTitleTopBar(
+fun TopBar(
     modifier: Modifier = Modifier,
     title: String = "",
+    navigationIcon: @Composable () -> Unit = {},
     actions: @Composable RowScope.() -> Unit = {},
 ) {
     TopAppBar(
         modifier = modifier
             .fillMaxWidth()
             .height(56.dp),
-        contentPadding = PaddingValues(end = 4.dp)
+        backgroundColor = YdsTheme.colors.bgElevated,
+        contentColor = YdsTheme.colors.textPrimary,
+        elevation = 0.dp,
+        contentPadding = PaddingValues(horizontal = 4.dp)
     ) {
         Box(
-            modifier = Modifier
-                .fillMaxWidth()
+            modifier = Modifier.fillMaxWidth()
         ) {
-            YdsText(
+            Row(
+                verticalAlignment = Alignment.CenterVertically, modifier = Modifier
+                    .fillMaxHeight()
+                    .align(Alignment.TopStart)
+                    .padding(0.dp)
+            ) {
+                navigationIcon()
+            }
+
+            Row(
                 modifier = Modifier
-                    .padding(16.dp, 17.dp, 16.dp, 8.dp)
-                    .wrapContentHeight(),
-                text = title,
-                style = YdsTheme.typography.title2,
-                color = YdsTheme.colors.textPrimary
-            )
+                    .fillMaxHeight()
+                    .padding(0.dp)
+                    .align(Alignment.TopCenter),
+                verticalAlignment = Alignment.CenterVertically,
+            ) {
+                YdsText(
+                    text = title,
+                    style = YdsTheme.typography.subTitle2,
+                    color = YdsTheme.colors.textPrimary
+                )
+            }
 
             Row(
                 modifier = Modifier
@@ -59,26 +75,28 @@ fun SingleTitleTopBar(
     }
 }
 
-@Preview(name = "SingleTitleTopBar")
+@Preview(name = "TopBar")
 @Composable
-private fun PreviewSingleTitleTopBar() {
+private fun PreviewTopBar() {
     YdsTheme {
         YdsScaffold(
             topBar = {
-                SingleTitleTopBar(
+                TopBar(
                     title = "타이틀",
+                    navigationIcon = {
+                        TopBarButton(
+                            icon = R.drawable.ic_arrow_left_line,
+                            isDisabled = false,
+                        )
+                    },
                     actions = {
                         TopBarButton(
-                            icon = R.drawable.ic_ground_filled,
-                            isDisabled = false
+                            icon = R.drawable.ic_bell_line,
+                            isDisabled = false,
                         )
                         TopBarButton(
-                            icon = R.drawable.ic_ground_filled,
-                            isDisabled = false
-                        )
-                        TopBarButton(
-                            icon = R.drawable.ic_ground_filled,
-                            isDisabled = false
+                            icon = R.drawable.ic_search_line,
+                            isDisabled = false,
                         )
                     }
                 )
